requirements.txt
python-dotenv 
requests 
bs4
python-dotenv
dnspython
markdown 
pdfkit
fastmcp<3
starlette
pytest